requirements.txt
Flask==3.1.0
pillow
requests
python-dotenv